refactor(projects): modularize project handling logic
- Extract project action handling into useProjectActions hook
- Create useAddProjectWorkflow for adding projects
- Implement useProjectDropImport for handling drag-and-drop
- Introduce useProjectsSort for managing project sorting state
- Clean up ProjectsView component for better readability and maintainability
